resolvedTickets added

+++ database/migrations/it_employee/2024_04_01_072726_add_resolved_tickets_in_it_employees_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('it_employees', function (Blueprint $table) {
-            $table->tinyInteger('resolved_tickets')->nullable();
+            $table->mediumInteger('resolved_tickets')->nullable();
         });
     }
